% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmean.R
\name{absmean}
\alias{absmean}
\title{Mean magnitude}
\usage{
absmean(x)
}
\arguments{
\item{x}{A numeric vector,}
}
\value{
A scalar holding the mean magnitude.
}
\description{
Compute the arithmetic mean of the absolute values.
}
