% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occuring.R
\name{also_occuring}
\alias{also_occuring}
\alias{only_occuring}
\alias{not_occuring}
\title{Weave schedules together}
\usage{
also_occuring(elem_1, elem_2)

only_occuring(elem_1, elem_2)

not_occuring(elem_1, elem_2 = NULL)
}
\description{
Joins schedules together.
}
\details{
Details ...
}
\examples{
on_christmas <- only_occuring(on_mday(25), in_month("Dec"))

on_christmas(as.Date("2000-12-25"))
on_christmas(as.Date("2000-12-26"))

schedule(on_christmas, from = 2000, to = 2001)

on_new_years_day <- on_yday(1)

on_public_holidays <- also_occuring(on_new_years_day, on_christmas)

on_public_holidays(as.Date("2000-12-25"))
on_public_holidays(as.Date("2000-12-27"))
on_public_holidays(as.Date("2001-01-01"))

schedule(on_public_holidays, from = 2000, to = 2001)


}
