% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurring.R
\name{also_occurring}
\alias{also_occurring}
\alias{only_occurring}
\title{Weave schedules together}
\usage{
also_occurring(x, y)

only_occurring(x, y)
}
\arguments{
\item{x, y}{Schedules to weave together.}
}
\value{
A schedule of events determined by the input schedules and rules
of the function used.
}
\description{
These functions combine schedules by taking two of them as inputs and
returning a single combined schedule as output.
}
\details{
\code{also_occurring()} returns a schedule of events which includes all those
present in the first schedule (\code{x}) and all those present in the second
schedule (\code{y}). The results is the \strong{union} of the two schedules.

\code{only_occurring()} returns a schedule of events which includes only those
present in both the first schedule (\code{x}) and in the second schedule
(\code{y}). The result is the \strong{intersection} of the two schedules.
}
\examples{
on_christmas <- only_occurring(on_mday(25), in_month("Dec"))

schedule(on_christmas, from = 2000, to = 2004)

on_new_years_day <- on_yday(1)

on_public_holidays <- also_occurring(on_new_years_day, on_christmas)

schedule(on_public_holidays, from = 2000, to = 2004)
}
