% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_temp.R
\name{as_temp}
\alias{as_temp}
\title{Generate the data structure for asreml-w}
\usage{
as_temp(dat)
}
\arguments{
\item{dat}{Data fram that used to be analysised.}
}
\value{
The structure of the data.
}
\description{
Generate the data structure for asreml-w
}
\examples{
library(asreml)
data(harvey)
as_temp(harvey)
}
