% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPhen.R
\name{checkPhen}
\alias{checkPhen}
\title{Check phenotypic Information}
\usage{
checkPhen(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
data depurated
}
\description{
\code{checkPhen} returns phenotypic information cleaned
}
\examples{
# library(sommer)
# data(DT_cpdata)
# phen <- DT_cpdata

# checkPhen(phen)

}
\author{
Johan Aparicio, \email{j.aparicio@cgiar.org}
}
