% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossGP.R
\name{crossGP}
\alias{crossGP}
\title{Cross validation}
\usage{
crossGP(geno, samp, phen, prior, niter = 50, testporc = 0.3, traits = NULL)
}
\arguments{
\item{geno}{The name of the file which the genotypic-data are to be read from or a matrix in the r-environment. Data coded as (-1,0,1)}

\item{samp}{The name of the file which the genotypes are to be read from or a vector  with the genotype names}

\item{phen}{The name of the file which the phenotypic-data are to be read from or a data.frame in the r-environment.}

\item{prior}{A string vector with some of those models c("ASReml", "RKHS", "sommer", "BRR", "BayesA", "BayesB", "BayesC", "BLasso")}

\item{niter}{A numeric from 0 to 100}

\item{testporc}{A numeric from 0 to 1}

\item{traits}{A string vector with the traits names}
}
\value{
list with two elements
}
\description{
\code{crossGP} returns the cross validation usign differents methodologies as implemented in BGLR, ASReml and sommer package.
}
\examples{

# library(sommer)
# data(DT_cpdata)
#
# geno <- GT_cpdata
# samp <- rownames(GT_cpdata)
# phen <- DT_cpdata
#
# crossGP(geno,samp,phen,prior = "sommer", niter=2,testporc = 0.3,traits = names(phen)[5])

#-----------------------

# geno <- "D:/OneDrive - CGIAR/2020/imputed_rrBLUP.in"
# samp <- "D:/OneDrive - CGIAR/2020/imputed_rrBLUP_samples.txt"
# phen <- "D:/OneDrive - CGIAR/2020/Phenotypic_Analysis.csv"
#
# crossGP(geno,samp,phen,prior = "sommer", niter=2,testporc = 0.3,traits = "Pal13C_drt")

}
\author{
Johan Aparicio, \email{j.aparicio@cgiar.org}
}
