% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_asremlw_H.R
\name{cv_asremlw_H}
\alias{cv_asremlw_H}
\title{Get the cv-result that run asremlw in the working directory}
\usage{
cv_asremlw_H(dat, Gmatrix, seed = 123)
}
\arguments{
\item{dat}{contins two columns: ID and y, that the ginv and the id.csv should}

\item{Gmatrix}{That is G-matrix, build by SNP marker, can be A.mat in sommer}

\item{seed}{set the seed number, default is 123}
}
\value{
result
}
\description{
Get the cv-result that run asremlw in the working directory
}
\examples{
1+1
}
