% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_asremlw_inv.R
\name{cv_asremlw_inv}
\alias{cv_asremlw_inv}
\title{Get the cv-result that run asremlw in the working directory}
\usage{
cv_asremlw_inv(dat, Hinv, seed = 123)
}
\arguments{
\item{dat}{contins two columns: ID and y, that the ginv and the id.csv should}

\item{Hinv}{That is Hinv-matrix, is inverse of H-matrix}

\item{seed}{set the seed number, default is 123}
}
\value{
result
}
\description{
Get the cv-result that run asremlw in the working directory
}
\examples{
1+1
}
