% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_cal.R
\name{cv_cal}
\alias{cv_cal}
\title{Build the H-matrix from G-inverse-matrix and pedigree-full and pedigree-genotype, It calculate alpha and bera based the function of the matrix G_22 and A_22}
\usage{
cv_cal(dat.csv, h2 = 1)
}
\arguments{
\item{M012}{It is the matrix which has rownames(ID), the first column is SNP infor, the SNP is 012 format.}

\item{ped_full}{It contains the full pedigree, it has three columns:ID,Sire,Dam}
}
\value{
The H-matrix form the formula
}
\description{
Build the H-matrix from G-inverse-matrix and pedigree-full and pedigree-genotype, It calculate alpha and bera based the function of the matrix G_22 and A_22
}
\examples{
library(MASS)
animal  <- 13:26
}
