% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_generater.R
\name{cv_generater}
\alias{cv_generater}
\title{Generate the K-fold number in cross-validation}
\usage{
cv_generater(y, k = n, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{The length of the y variate}

\item{k}{the k folder, is a number}
}
\value{
K-subset of the length
}
\description{
Generate the K-fold number in cross-validation
}
\examples{
dat <- data.frame(ID=paste("ID",151:160,sep=""),y=rnorm(10))
dat
index <- cv_generater(1:length(dat$y),k=3)
dat$y1 = dat$y2 = dat$y3 =dat$y;
index
for(i in 1:length(index)){
  vv <- dat$ID[index[[i]]]
  vv <- as.character(vv)
  dat[match(vv,dat$ID),i+2] <- NA
}
dat
index

}
