% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gblup_cv_model.R
\name{gblup_cv_model}
\alias{gblup_cv_model}
\title{Get the arrucacy and other values when run crossvalidation}
\usage{
gblup_cv_model(pheno_data, inv)
}
\arguments{
\item{pheno_data}{contins two columns: ID and y}

\item{inv}{contains three columns that is ginv or ainv: row, col, inverse_values}
}
\value{
result
}
\description{
Get the arrucacy and other values when run crossvalidation
}
\examples{

}
