% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_asreml_template.R
\name{Generate_asreml_template}
\alias{Generate_asreml_template}
\title{Generate the dat.as for asreml-w}
\usage{
Generate_asreml_template(dat)
}
\arguments{
\item{dat}{Data.frame structure}
}
\value{
Return the file in default workong directory named:Model-dave.as
}
\description{
Generate the dat.as for asreml-w
}
\examples{
library(asreml)
data(harvey)
Generate_asreml_template(harvey)


}
