% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGS.R
\name{getGS}
\alias{getGS}
\title{Gradient synchronization (GS)}
\usage{
getGS(t, x, y)
}
\arguments{
\item{t}{A vector of support grid in a strictly increasing order. 
If functions are not supported on \eqn{[0,1]}, \code{t} will be normalized 
to \eqn{[0,1]} by \code{(t-min(t)) / diff(range(t))}.}

\item{x, y}{The values of the two functions evaluated on \code{t}.}
}
\value{
A scalar holding the gradient synchronization.
}
\description{
Estimate the gradient synchronization (GS) for 
a pair of curves observed on the same grid.
}
