% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSF.R
\name{getGSF}
\alias{getGSF}
\title{Gradient synchronization fluctuation (GSF)}
\usage{
getGSF(t, x, y)
}
\arguments{
\item{t}{A vector of support grid in a strictly increasing order.}

\item{x, y}{The values of the two functions evaluated on \code{t}.}
}
\value{
A scalar holding the gradient synchronization fluctuation.
}
\description{
Estimate the gradient synchronization fluctuation (GSF) for 
a pair of curves observed on the same grid.
}
