% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happen.R
\name{happen}
\alias{happen}
\title{Does a date fall on a schedule?}
\usage{
happen(schedule, date)
}
\arguments{
\item{schedule}{A schedule object}

\item{date}{a date-time object}
}
\value{
TRUE or FALSE depending on whether the date falls on the schedule.
}
\description{
Does a date fall on a schedule?
}
\examples{
my_dates <- seq.Date(from = as.Date("2000-01-01"),
                     to = as.Date("2000-01-10"),
                     by = "1 day")

on_saturday <- on_wday("Sat")
on_sunday <- on_wday("Sun")

happen(on_saturday, my_dates)
happen(on_sunday, my_dates)
happen(on_weekend(), my_dates)
}
\keyword{date,}
\keyword{schedule}
