% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinv_matrix.R
\name{hinv_matrix}
\alias{hinv_matrix}
\title{Build the H-inverse-matrix from G-inverse-matrix and pedigree-full and pedigree-genotype#'}
\usage{
hinv_matrix(G_mat, ped_full, ped_geno)
}
\arguments{
\item{G_mat}{It is the matrix which has rownames and colnames(ID)}

\item{ped_full}{It contains the full pedigree, it has three columns:ID,Sire,Dam}

\item{ped_geno}{It contains the pedigree that has the genotype, it is a part of the ped_full pedigree}
}
\value{
The H-inverse-matrix form the formula
}
\description{
Build the H-inverse-matrix from G-inverse-matrix and pedigree-full and pedigree-genotype#'
}
\examples{
1+1

}
