% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter.R
\name{in_quarter}
\alias{in_quarter}
\alias{in_semester}
\title{Specify the quarter(s) of a schedule}
\usage{
in_quarter(..., fiscal_start = 1)

in_semester(x)
}
\arguments{
\item{...}{A numeric vector of quarter specifications.}

\item{fiscal_start}{For fiscal quarters, a numeric indicating the starting
month of a fiscal year. Defaults to 1 (January).}

\item{x}{Numeric specifying which semester the events occur in (1 or 2).}
}
\value{
A schedule object.
}
\description{
Creates a schedule of events occurring in the quarters or semester specified.
}
\examples{
 my_dates <- seq.Date(from = as.Date("2000-01-01"),
                     to = as.Date("2000-12-01"),
                     by = "1 month")

happen(in_quarter(1), my_dates)

happen(in_quarter(1, 3), my_dates)

happen(in_quarter(1:3), my_dates)

happen(in_quarter(1, 3, fiscal_start = 2), my_dates)

happen(in_semester(2), my_dates)
}
\keyword{date,}
\keyword{quarter,}
\keyword{schedule}
\keyword{semester,}
