% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{in_week}
\alias{in_week}
\alias{in_isoweek}
\title{Specify the weeks of a schedule}
\usage{
in_week(x, ...)

in_isoweek(x, ...)
}
\arguments{
\item{...}{a numeric vector of week elements.}
}
\value{
A schedule of events occuring in the weeks specified.
}
\description{
\code{in_week()} returns the number of complete seven day periods
that have occurred between the date and January 1st, plus one.

The \code{in_isoweek()} function returns the week as it would appear in the ISO 8601
system, which uses a reoccurring leap week.
}
\details{
Creates a schedule of events occuring in the weeks specified.
}
\examples{
my_dates <- c(as.Date("2000-01-01"),
              as.Date("2000-02-01"),
              as.Date("2000-03-01"),
              as.Date("2000-04-01"))

my_dates

in_january <- in_month("January")
in_january(my_dates)

in_february <- in_month("Feb")
in_february(my_dates)

in_march <- in_month(3)
in_march(my_dates)

in_jan_or_feb <- in_month("Jan", "Feb")
in_jan_or_feb(my_dates)

## You can even mix your month specs.
in_jan_feb_mar <- in_month("January", "Feb", 3)
in_jan_feb_mar(my_dates)

## invalid inputs will produce an immediate error
\dontrun{
in_january <- in_month("Janu")
in_january <- in_month(13)}
}
\keyword{date,}
\keyword{month,}
\keyword{scedule}
