% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{in_year}
\alias{in_year}
\alias{in_isoyear}
\alias{in_epiyear}
\title{Specify the year(s) of a schedule}
\usage{
in_year(...)

in_isoyear(...)

in_epiyear(...)
}
\arguments{
\item{...}{A numeric vector of years.}
}
\value{
A schedule object.
}
\description{
Creates a schedule of events occurring in the year(s) specified.
}
\details{
The type of year is determined by the function used. These year types are
built atop their \link[lubridate:year]{definitions} in the \code{lubridate}
package, where \code{year()} is the commonly understood definition of
a year. The other definitions from the \code{lubridate} package are quoted here:
\itemize{
\item \code{isoyear()} returns years according to the ISO 8601 week calendar.
\item \code{epiyear()} returns years according to the epidemilogical week calendars.
}

The \code{in_year()} function will encode limits on the output schedule.
For more details see the
\href{https://jameslairdsmith.github.io/scheduler/articles/understanding-schedule-limits.html}{vignette}
on understanding schedule limits.
}
\examples{
my_dates <- seq(from = as.Date("2000-01-01"),
                to = as.Date("2005-01-01"),
                by = "1 year")

is_occurring(my_dates, in_year(2000))

is_occurring(my_dates, in_year(2000, 2002))

is_occurring(my_dates, in_year(2000:2002))

is_occurring(my_dates, in_year(2000:2002, 2005))

is_occurring(my_dates, in_isoyear(2004))

head(schedule(in_year(2000), 20))
}
\keyword{date,}
\keyword{schedule}
\keyword{year,}
