% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_occurring}
\alias{is_occurring}
\title{Does a date fall on a schedule?}
\usage{
is_occurring(date, schedule)
}
\arguments{
\item{date}{a date-time object}

\item{schedule}{A schedule object}
}
\value{
TRUE or FALSE depending on whether the date falls on the schedule.
}
\description{
Does a date fall on a schedule?
}
\examples{
my_dates <- seq.Date(from = as.Date("2000-01-01"),
                     to = as.Date("2000-01-10"),
                     by = "1 day")

on_saturday <- on_wday("Sat")
on_sunday <- on_wday("Sun")

is_occurring(my_dates, on_saturday)
is_occurring(my_dates, on_sunday)
is_occurring(my_dates, on_weekend())
}
\keyword{date,}
\keyword{schedule}
