% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easter.R
\name{on_easter_sunday}
\alias{on_easter_sunday}
\alias{on_good_friday}
\alias{on_easter_monday}
\title{Schedule dates during Easter}
\usage{
on_easter_sunday()

on_good_friday()

on_easter_monday()
}
\value{
A schedule object.
}
\description{
Creates a schedule of events occuring on the various Easter holidays.
}
\examples{
schedule_days(on_easter_sunday(), from = 2010, to = 2020)

schedule_days(on_good_friday(), from = 2010, to = 2020)

schedule_days(on_easter_saturday(), from = 2010, to = 2020)
}
\keyword{easter,}
\keyword{schedule}
