% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every.R
\name{on_every_nth}
\alias{on_every_nth}
\alias{on_every}
\title{Schedule events incrementally}
\usage{
on_every_nth(n, unit, starting, inclusive = TRUE, backdated = FALSE)

on_every(unit, starting, inclusive = TRUE, backdated = FALSE)
}
\arguments{
\item{n}{A single integer specifying the increment of the events to create.}

\item{unit}{The type of increment. Can be either:
\itemize{
\item A \code{lubridate} period object eg. "months".
\item A schedule object.
}}

\item{starting}{The first event of the series. Can be a date or datetime.}

\item{inclusive}{Logical indicating whether the \code{starting} date should
be included in the resulting schedule.}

\item{backdated}{Logical indicating whether dates prior the \code{starting}
date should be included in the resulting schedule.}
}
\value{
A schedule object.
}
\description{
Create a schedule of events occuring in increments from a fixed date.
}
