% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule_days}
\alias{schedule_hours}
\title{Get the events of a schedule}
\usage{
schedule(x, from = NULL, to = NULL, during = NULL, every = "1 day",
  ...)

schedule_days(x, from = NULL, to = NULL, during = NULL, ...)

schedule_hours(x, from = NULL, to = NULL, during = NULL, ...)
}
\arguments{
\item{...}{A numeric vector of year elements.}
}
\value{
A date vector
}
\description{
Get the events from a schedule object.
}
\details{
R
}
\examples{
on_first_day_month <- on_mday(1)

schedule(on_first_day_month, during = 2000)
}
\keyword{schedule}
