% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse2mat.R
\name{sparse2mat}
\alias{sparse2mat}
\title{Generate inv to mat, use DMU formart to asreml format}
\usage{
sparse2mat(x)
}
\arguments{
\item{x}{sparse inv that have three column:id1 id2 value}
}
\value{
matrix
}
\description{
Generate inv to mat, use DMU formart to asreml format
}
\examples{
library(GS)
dmu_inv <- data.frame(ID1 = rep(1:3,each=3),ID2 = rep(1:3,3),value=rnorm(9)+5)
dmu_inv
asr_mat <- sparse2mat(dmu_inv)
asr_inv <- write_mat_to_inv(asr_mat,"none")
asr_inv
}
