% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mat_to_inv.R
\name{write_mat_to_inv}
\alias{write_mat_to_inv}
\title{Change the G-matrix ginv that contains three columns format}
\usage{
write_mat_to_inv(x, type = c("ginv", "inv", "none"), digits = 10)
}
\arguments{
\item{x}{A matrix that is G-matrix}
}
\value{
Three column that contains row, column, and the value of a matrix
}
\description{
Change the G-matrix ginv that contains three columns format
}
\examples{
rMinv <- matrix(1:25,5,5)
rMinv
res <- write_mat_to_inv(rMinv,type = "none")
res
}
