% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{after}
\alias{after}
\alias{before}
\title{Specify the start and end of a schedule}
\usage{
after(x, within_given = NULL)

before(x, within_given = NULL)
}
\arguments{
\item{x}{Either a date object or a schedule.}

\item{within_given}{A date accessor function. Required only when \code{x} is a
schedule. Puts a limit on how far into the future the schedule will
extend. Used to avoid an infinite cycle.}
}
\value{
A schedule of events occuring after or before the events specified.
}
\description{
Creates a schedule of events occuring before or after a certain date or
another scheduled event within a period.
}
\details{
When \code{x} is a date.

When \code{x} is a schedule...

When \code{x} is a date, it saves having to...
}
\examples{
on_christmas <- on_mday(25) \%>\% only_occuring(in_month("Dec"))

after_christmas <- after(on_christmas, within_given = lubridate::year)

schedule(after_christmas, from = 2000, to = 2001)
}
\keyword{after,}
\keyword{before,}
\keyword{date,}
\keyword{schedule}
