% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{after}
\alias{after}
\alias{before}
\alias{in_between}
\title{Specify the start and end of a schedule}
\usage{
after(start_event, within_given = NULL)

before(end_event, within_given = NULL)

in_between(start_event, end_event, within_given = NULL)
}
\arguments{
\item{start_event, end_event}{The start and/or end events of the schedule.
For each can be either be a date object or a schedule.}

\item{within_given}{A date accessor function. Required only when either
\code{start_date} or \code{end_date} are schedules. Puts a limit on how far into the
future the schedule will extend. Used to avoid an infinite cycle.}
}
\value{
A schedule of events occuring before and/or after the events specified.
}
\description{
Creates a schedule of events occuring before and/or after a certain date or
scheduled event.
}
\details{
When \code{x} is a date.

When \code{x} is a year.

When \code{x} is a schedule...

When \code{x} is a date, it saves having to...
}
\examples{
on_christmas <- on_mday(25) \%>\% only_occuring(in_month("Dec"))

after_christmas <- after(on_christmas, within_given = lubridate::year)

schedule(after_christmas, from = 2000, to = 2001)

on_new_years_eve <- on_mday(31) \%>\% only_occuring(in_month("Dec"))

in_between_christmas_and_new_years_eve <-
    in_between(on_christmas,
               on_new_years_eve,
               within_given = lubridate::year)

schedule(in_between_christmas_and_new_years_eve, from = 2000, to = 2001)

}
\keyword{after,}
\keyword{before,}
\keyword{date,}
\keyword{schedule}
