% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{after}
\alias{after}
\alias{before}
\title{Create a schedule of events occuring before or after other events}
\usage{
after(x, within_given = NULL)

before(x, within_given = NULL)
}
\arguments{
\item{x}{Either a date object or a schedule.}

\item{within_given}{Required when}
}
\value{
A schedule of events occuring after or before the events specified.
}
\description{
Creates a schedule of events occuring before or after a certain date or
another scheduled event within a certain period.
}
\details{
When \code{x} is a date.

When \code{x} is a schedule...

When \code{x} is a date, it saves having to...
}
\examples{
x <- 1
x
}
\keyword{after,}
\keyword{before,}
\keyword{date,}
\keyword{schedule}
