% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossGP.R
\name{crossGP}
\alias{crossGP}
\title{Cross validation}
\usage{
crossGP(geno, samp, phen, prior, niter = 50, testporc = 0.3, traits = NULL)
}
\arguments{
\item{geno}{A string}

\item{samp}{A string}

\item{phen}{A string}

\item{prior}{A string vector c("ASReml", "RKHS", "sommer", "BRR", "BayesA", "BayesB", "BayesC", "BLasso")}

\item{niter}{A numeric}

\item{testporc}{A numeric}

\item{traits}{A string vector}
}
\value{
list with two elements
}
\description{
\code{crossGP} returns the cross validation usign differents methodologies as implemented in GBLR, ASReml and sommer package.
}
\examples{

\dontrun{
 geno <- "imputed_rrBLUP.in"
 samp <- "imputed_rrBLUP_samples.txt"
 phen <- "Phenotypic_Analysis.csv"

 prior <- c("ASReml", "RKHS", "sommer")

 out_table <- crossGP(geno, samp, phen, prior, niter = 2, testporc = 0.3)
}
}
