% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{in_month}
\alias{in_month}
\title{Create a month schedule}
\usage{
in_month(...)
}
\arguments{
\item{...}{months in which the schedule occurs.}
}
\value{
A month date element
}
\description{
Allows users to create an abstract object representing all dates taking
place within a
specific month. Users can specify which month using:
}
\details{
1) The month name (eg. "January"),
2) The month's abbreviation (eg. "Jan") or
3) the month number within the year (eg. 1).
}
\examples{
library(magrittr)
library(lubridate)

January <- in_month("Jan")

test_date(dmy("31/01/1990"), January)
}
\keyword{date,}
\keyword{month,}
\keyword{scedule}
