% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{in_month}
\alias{in_month}
\title{Specify the months of a schedule}
\usage{
in_month(...)
}
\arguments{
\item{...}{a character or numeric vector of month elements.}
}
\value{
A schedule of events occuring in the months specified.
}
\description{
Creates a schedule of events occuring in the months specified.
}
\details{
Months can be specified using their names (eg. "January"),
abbreviations (eg. "Jan") or integers (eg. 1).
}
\examples{
my_dates <- c(as.Date("2000-01-01"),
              as.Date("2000-02-01"),
              as.Date("2000-03-01"),
              as.Date("2000-04-01"))

my_dates

in_january <- in_month("January")
in_january(my_dates)

in_february <- in_month("Feb")
in_february(my_dates)

in_march <- in_month(3)
in_march(my_dates)

in_jan_or_feb <- in_month("Jan", "Feb")
in_jan_or_feb(my_dates)

## You can even mix your month specs, although I don't recommend it
in_jan_feb_mar <- in_month("January", "Feb", 3)
in_jan_feb_mar(my_dates)

## invalid inputs will produce an immediate error
\dontrun{
in_january <- in_month("Janu")
in_january <- in_month(0)}
}
\keyword{date,}
\keyword{month,}
\keyword{scedule}
