% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{in_month}
\alias{in_month}
\title{Specify the month(s) of a schedule}
\usage{
in_month(...)
}
\arguments{
\item{...}{A character or numeric vector of month elements.}
}
\value{
A schedule object.
}
\description{
Creates a schedule of events occurring in the months specified.
}
\details{
Months can be specified using their names (eg. "January"),
abbreviations (eg. "Jan") or integers (eg. 1).

Multiple months can be specified in the same function call. For example:
\code{in_month(1, 3)} produces a schedule of events occurring in January and
March. The same thing is achieved by \code{in_month("Jan", "Mar")}.
Similarly \code{in_month(1:3)} produces a schedule of events occurring in
January, February and March.
}
\examples{
my_dates <- seq.Date(as.Date("2000-01-01"),
                     as.Date("2000-04-01"),
                     by = "1 month")

is_occurring(my_dates, in_month("January"))

is_occurring(my_dates, in_month("Feb"))

is_occurring(my_dates, in_month(3))

is_occurring(my_dates, in_month("Jan", "Mar"))

is_occurring(my_dates, in_month(1:3))

## invalid inputs will produce an immediate error
\dontrun{
in_january <- in_month("Janu")
in_january <- in_month(0)}
}
\keyword{date,}
\keyword{month,}
\keyword{schedule}
