% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{in_month}
\alias{in_month}
\title{Create a month element}
\usage{
in_month(x, override_name_check = FALSE, ...)
}
\arguments{
\item{x}{The name, abbreviation or number of the month.}

\item{override_name_check}{Whether to allow month names and abbreviations
not found in in your default local (not recommended).}
}
\value{
A month date element
}
\description{
Allows users to create object that embody all dates taking place within a
specific month. Users can specify which month using either 1) The month
name (eg. "January"), the month's abbreviation (eg. "Jan") or the month
number within the year (eg. 1).
}
\examples{
library(magrittr)
library(lubridate)

January <- in_month("Jan")

test_date.Date(dmy("31/01/1990"), January)
}
\keyword{date,}
\keyword{month,}
\keyword{scedule}
