% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter.R
\name{in_quarter}
\alias{in_quarter}
\alias{in_semester}
\title{Specify the quarter(s) of a schedule}
\usage{
in_quarter(..., fiscal_start = 1)

in_semester(x)
}
\arguments{
\item{...}{A numeric vector of quarter specifications.}

\item{fiscal_start}{Numeric indicating the starting month of a fiscal year.
Defaults to 1 (January).}

\item{x}{Numeric specifying which semester the events occur in (1 or 2).}
}
\value{
A schedule of events occurring in the quarters or semester specified.
}
\description{
Creates a schedule of events occurring in the quarters or semester specified.
}
\examples{
 my_dates <- seq.Date(from = as.Date("2000-01-01"),
                     to = as.Date("2000-12-01"),
                     by = "1 month")

in_q1 <- in_quarter(1)
in_fiscal_q1 <- in_quarter(1, fiscal_start = 2)

in_q1(my_dates)
in_fiscal_q1(my_dates)

in_q1_or_q2 <- in_quarter(1, 2)
in_fiscal_q1_or_q2 <- in_quarter(1, 2, fiscal_start = 2)

in_q1_or_q2(my_dates)
in_fiscal_q1_or_q2(my_dates)

in_second_semester <- in_semester(2)

in_second_semester(my_dates)
}
\keyword{date,}
\keyword{quarter,}
\keyword{schedule}
\keyword{semester,}
