% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{in_week}
\alias{in_week}
\alias{in_isoweek}
\alias{in_epiweek}
\title{Specify the week(s) of a schedule}
\usage{
in_week(...)

in_isoweek(...)

in_epiweek(...)
}
\arguments{
\item{...}{a numeric vector of week specifications.}
}
\value{
A schedule of events occurring in the weeks specified.
}
\description{
Creates a schedule of events occurring in the weeks specified.
}
\details{
The type of week is determined by the function used. These week types are
built atop their definitions from the \code{lubridate} package, which are quoted
here:
\itemize{
\item \code{week()} returns the number of complete seven day periods that have
occurred between the date and January 1st, plus one.
}

\itemize{
\item \code{isoweek()} returns the week as it would appear in the ISO 8601
system, which uses a reoccurring leap week.
}

\itemize{
\item \code{epiweek()} is the US CDC version of epidemiological week. It
follows same rules as \code{isoweek()} but starts on Sunday. In other parts of
the world the convention is to start epidemiological weeks on Monday,
which is the same as isoweek.
}
}
\examples{
my_dates <- seq.Date(from = as.Date("2000-01-01"),
                     to = as.Date("2000-02-01"),
                     by = "1 week")
my_dates

in_first_week_year <- in_week(1)
in_first_week_year(my_dates)

in_first_or_third_week_year <- in_week(1, 3)
in_first_or_third_week_year(my_dates)

in_first_three_weeks_year <- in_week(1:3)
in_first_three_weeks_year(my_dates)

in_first_isoweek_year <- in_isoweek(1)
in_first_isoweek_year(my_dates)

## invalid inputs will produce an immediate error:
\dontrun{
in_week(0)
in_week(54)
in_week(1.5)}
}
\keyword{date,}
\keyword{scedule}
\keyword{week,}
