% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{in_year}
\alias{in_year}
\title{Specify the year(s) of a schedule}
\usage{
in_year(...)
}
\arguments{
\item{...}{A numeric vector of year elements.}
}
\value{
A schedule of events occuring in the year(s) specified.
}
\description{
Creates a schedule of events occuring in the year(s) specified.
}
\details{
The \code{in_year()} function will retain information on the limits of
a schedule. If a schedule occurs only within a specified year or years, we
can infer the start date and end date of the schedule as being the first
day of a earliest year and the last day of the latest year respectively.
This means we do not need to specify these limits when passing the schedule
on to the \code{schedule()} function (see last of the examples below).
}
\examples{
my_dates <- seq(from = as.Date("2000-01-01"),
                to = as.Date("2004-01-01"),
                by = "1 year")

in_2000 <- in_year(2000)
in_2000(my_dates)

in_2000_or_2001 <- in_year(2000, 2001)
in_2000_or_2001(my_dates)

in_2000_or_2001_or_2002 <- in_year(2000:2002)
in_2000_or_2001_or_2002(my_dates)

head(schedule(in_2000), 20)
}
\keyword{date,}
\keyword{schedule}
\keyword{year,}
