% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_occurring}
\alias{is_occurring}
\title{Does a date fall on a schedule?}
\usage{
is_occurring(date, schedule)
}
\arguments{
\item{date}{a date-time object}

\item{schedule}{A schedule object}
}
\value{
A logical.
}
\description{
Tests whether a date falls on a schedule.
}
\examples{
on_saturdays <- on_wday("Sat")

is_occurring(as.Date("2000-01-01"), on_saturdays)
is_occurring(as.Date("2000-01-02"), on_saturdays)
}
\keyword{date,}
\keyword{schedule}
