% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{on_date}
\alias{on_date}
\title{Specify the date(s) of a schedule}
\usage{
on_date(...)
}
\arguments{
\item{...}{A vector of dates.}
}
\value{
A schedule of events occuring on the dates specified.
}
\description{
Creates a schedule of events occuring on the dates specified.
}
\details{
This function is best used when making changes to other schedules where
those changes fall on defined dates.

The function also encodes schedule limits in the output. For more
details see the \href{https://jameslairdsmith.github.io/scheduler/articles/understanding-schedule-limits.html}{vignette}
on understanding schedule limits.
}
\examples{

on_first_day_millennium <- on_date(as.Date("2000-01-01"))
on_first_or_second_day_millennium <- on_date(as.Date("2000-01-01"),
                                             as.Date("2000-01-02"))

on_first_day_millennium(as.Date("2000-01-01"))
on_first_day_millennium(as.Date("2000-01-02"))
on_first_or_second_day_millennium(as.Date("2000-01-01"))
on_first_or_second_day_millennium(as.Date("2000-01-02"))
on_first_or_second_day_millennium(as.Date("2000-01-03"))
}
\keyword{date,}
\keyword{schedule}
