% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day.R
\name{on_mday}
\alias{on_mday}
\alias{on_yday}
\alias{on_wday}
\alias{on_weekend}
\title{Specify the days of a schedule}
\usage{
on_mday(...)

on_yday(...)

on_wday(..., week_start = getOption("lubridate.week.start", 7))

on_weekend()
}
\arguments{
\item{...}{a numeric vector of day elements. In the case of \code{on_wday} the
elements can also be characters (see details below).}

\item{week_start}{You can use \code{lubridate}'s global option \code{lubridate.week.start} to set this
parameter globally}
}
\value{
A schedule of events occuring on the day types specified.
}
\description{
Creates a schedule of events occuring on the types of days specified.
}
\details{
\code{on_weekend()} is a convenience function for
\code{on_wday("Sat", "Sun")}
}
\keyword{date,}
\keyword{month,}
\keyword{scedule}
