% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{on_nth}
\alias{on_nth}
\alias{on_first}
\alias{on_second}
\alias{on_third}
\alias{on_fourth}
\alias{on_last}
\title{Schedule the nth events within a period}
\usage{
on_nth(n, x, within_given, ...)

on_first(x, within_given, ...)

on_second(x, within_given, ...)

on_third(x, within_given, ...)

on_fourth(x, within_given, ...)

on_last(x, within_given, ...)
}
\arguments{
\item{x, y}{Schedules to weave together.}
}
\value{
A schedule of events.
}
\description{
Get the nth event of a certain type within a given period.
}
\details{
R
}
\examples{
on_tuesday <- on_wday("Tue")

on_fourth_tuesday <- on_nth(4, on_tuesday, within_given = "month")

schedule(on_fourth_tuesday, during = 2000)

}
