% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_date}
\alias{test_date}
\title{Test whether a date falls on a schedule.}
\usage{
test_date(date, schedule, ...)
}
\arguments{
\item{date}{A date object}

\item{schedule}{A schedule object}
}
\description{
Deprecated. Instead use \code{is_occurring()}.
}
